refactor: decouple simple conv net from benchmark
